requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-pydenticon==0.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
opbeat==3.5.2
openapi-codec==1.3.1
Pillow==4.2.1
pydenticon==0.3
pytz==2017.2
requests==2.18.1
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.21.1
